✨ 타입 문제로 인하여 mock데이터와 연동되지않았던 데이터 연동
